refactor; add manual control, stateful resume
Rewrote most of the code with more descriptive variable names and extensive comments.  Added manual pump start/stop with `m,start` and `m,stop`.  Pausing the controller now keeps track of the pump and cycle state, so the system returns to the pump on/off state and its position in the on/off control cycle when the controller is resumed.